split out dataset selector test
